// src/types/machine.ts
export default interface Machine{
    actual_lights: string[]
    bases_added: number
    bases_used: number
    bs_color: string
    bs_side: string
    cs_operation: string
    cs_retrieved: boolean
    ds_order: number
    level: string
    mtype: string
    name: string
    rotation: number
    rs_ring_color: string
    rs_ring_colors: string[]
    state: string
    team: string
    type: string
    zone: string
}